Hack: Use serial number to generate wifi api
This commit changes the SSID of the Wifi AP to PREFIX-<SERIALNUMBER> in
accordance with Eltako wifi ap. This change is likely not compatible
with Matter. However we do not support commissioning using Wifi anyway.
Yet we need wifi to be available to configure the device by other means.
In the future we need to find a way to either enable the ap without
using Matter functions or provide proper commissioning over wifi.
The latter is likely the better way.